Latest from indi master

+++ libindi/drivers/auxiliary/gpusb.h
@@ -50,10 +50,10 @@ class GPUSB : public INDI::GuiderInterface, public INDI::DefaultDevice
 
     void TimerHit();
 
-    virtual IPState GuideNorth(float ms);
-    virtual IPState GuideSouth(float ms);
-    virtual IPState GuideEast(float ms);
-    virtual IPState GuideWest(float ms);
+    virtual IPState GuideNorth(uint32_t ms);
+    virtual IPState GuideSouth(uint32_t ms);
+    virtual IPState GuideEast(uint32_t ms);
+    virtual IPState GuideWest(uint32_t ms);
 
   private:
     float CalcWEPulseTimeLeft();
