+++ libindi/drivers/telescope/paramount.h
@@ -64,10 +64,10 @@ class Paramount : public INDI::Telescope, public INDI::GuiderInterface
     virtual bool SetDefaultPark() override;
 
     // Guiding
-    virtual IPState GuideNorth(float ms) override;
-    virtual IPState GuideSouth(float ms) override;
-    virtual IPState GuideEast(float ms) override;
-    virtual IPState GuideWest(float ms) override;
+    virtual IPState GuideNorth(uint32_t ms) override;
+    virtual IPState GuideSouth(uint32_t ms) override;
+    virtual IPState GuideEast(uint32_t ms) override;
+    virtual IPState GuideWest(uint32_t ms) override;
 
   private:
     void mountSim();
